requirements.txt
numpy>=1.20.0
matplotlib>=3.5.0
sympy>=1.10.0
pint>=0.20.0
jupyter

cadquery
python-dotenv
rhino
rhinoinside